apply axios and pagination

+++ src/app/api/articles/[id]/route.ts
@@ -2,6 +2,9 @@ import { articles } from "@/Utils/data";
 import { NextRequest, NextResponse } from "next/server";
 import { UpdateArticleDto } from "@/Utils/dtos";
 import prisma from "@/Utils/db";
+import { verifyToken } from "@/Utils/verifyToken";
+import { comma } from "postcss/lib/list";
+import { comment } from "postcss";
 
 interface Props {
     params: { id: string }
@@ -16,7 +19,23 @@ interface Props {
 
 export async function GET(request: NextRequest, { params }: Props) {
     try {
-        const article = await prisma.article.findUnique({ where: { id: parseInt(params.id) } });
+        const article = await prisma.article.findUnique({ 
+                where: { id: parseInt(params.id) },
+                include: {
+                    comments: {
+                        include: {
+                            user: {
+                                select: {
+                                    username: true
+                                }
+                            }
+                        },
+                        orderBy:{
+                            createdAt: 'desc'
+                        }
+                    }
+                }
+            });
         if (!article) {
             return NextResponse.json({ message: 'article not found' }, { status: 404 })
         }
@@ -31,11 +50,19 @@ export async function GET(request: NextRequest, { params }: Props) {
  * @method PUT
  * @route ~/api/articles/:id
  * @desc Update Article
- * @access public 
+ * @access private (only admin can update articles) 
  */
 
 export async function PUT(request: NextRequest, { params }: Props) {
-    const article = await prisma.article.findUnique({
+    try {
+        const user = verifyToken(request);
+        if (user === null || user.isAdmin === false) {
+            return NextResponse.json(
+                {message: 'only admin, access denied'},
+                {status: 403}
+            )
+        }
+        const article = await prisma.article.findUnique({
         where: { id: parseInt(params.id) }
     });
 
@@ -52,28 +79,50 @@ export async function PUT(request: NextRequest, { params }: Props) {
         }
     })
     return NextResponse.json({ 'message': 'article updated' }, { status: 200 })
+} catch (error) {
+    return NextResponse.json(
+        { message: "internal server error" },
+        { status: 500 }
+    )
+}
 }
 
 /**
  * @method DELETE
  * @route ~/api/articles/:id
  * @desc Delete Article
- * @access public 
+ * @access private (only admin can delete article) 
  */
 
 export async function DELETE(request: NextRequest, { params }: Props) {
     try {
+        const user = verifyToken(request);
+        if (user === null || user.isAdmin === false) {
+            return NextResponse.json(
+                {message: 'only admin, access denied'},
+                {status: 403}
+            )
+        }
         const article = await prisma.article.findUnique({
-            where: { id: parseInt(params.id) }
+            where: { id: parseInt(params.id) },
+            include: {comments: true}
         });
 
         if (!article) {
             return NextResponse.json({ 'message': 'article not found' }, { status: 404 })
         }
 
+        //deleting the article
         await prisma.article.delete({
             where: { id: parseInt(params.id) }
+        });
+
+        //deleting the comments that belong to this article
+        const commentsIds: number[] = article?.comments.map(comment => comment.id)
+        await prisma.comment.deleteMany({
+            where: {id: {in: commentsIds}}
         })
+
         return NextResponse.json({ 'message': 'article deleted' }, { status: 200 })
     } catch (error) {
         return NextResponse.json(
